EDIT: moved helpers to src folder
